src/uav_web_control/requirements.txt
Flask>=2.0
PyYAML>=5.4
